Add workflow to rename repository upon initialization
This commit adds a GitHub Actions workflow to rename a repository upon initialization. The workflow checks if the repository is not feelpp-project, initializes it by renaming it to its own name, commits the changes with a message that includes the new name, pushes the changes with a force flag, and finally disables the workflow.

automatically rename project when used as template #91