add support for python bindings #122
Add support for python bindings and update installation process

This commit adds support for python bindings by adding a python library module, `_laplacian`, and linking it to the `pybind11` headers and `feelpp` library. The `VERSION_INFO` is set to the `PROJECT_VERSION` for the `_laplacian` module. The installation process is updated to install the `_laplacian` module if `SKBUILD_PROJECT_NAME` is present, otherwise, a python module `laplacian` is added and relevant files are installed to the appropriate directories.